refactor: refactored components for better modularity
