+++ src/features/inventory/component/inventory-form/stock-slider.tsx
@@ -1,7 +1,7 @@
-import { useState, useEffect } from "react";
-import { Label } from "components/ui/label";
-import { Input } from "components/ui/input";
-import { cn } from "lib/utils";
+import { useState, useEffect } from 'react';
+import { Label } from 'components/ui/label';
+import { Input } from 'components/ui/input';
+import { cn } from 'lib/utils';
 
 interface StockSliderProps {
   label?: string;
@@ -15,7 +15,7 @@ interface StockSliderProps {
 }
 
 export function StockSlider({
-  label = "Stock",
+  label = 'Stock',
   value = 0,
   min = 0,
   max = 1000,
@@ -43,7 +43,7 @@ export function StockSlider({
     const newValue = e.target.value;
     setInputValue(newValue);
 
-    if (newValue !== "" && !isNaN(Number(newValue))) {
+    if (newValue !== '' && !isNaN(Number(newValue))) {
       const numValue = Number.parseInt(newValue);
       const boundedValue = Math.min(Math.max(numValue, min), max);
       setSliderValue(boundedValue);
@@ -52,7 +52,7 @@ export function StockSlider({
   };
 
   const handleInputBlur = () => {
-    if (inputValue === "" || isNaN(Number(inputValue))) {
+    if (inputValue === '' || isNaN(Number(inputValue))) {
       setInputValue(sliderValue.toString());
     } else {
       const numValue = Number.parseInt(inputValue);
@@ -64,7 +64,7 @@ export function StockSlider({
   };
 
   return (
-    <div className={cn("flex flex-col gap-2", className)}>
+    <div className={cn('flex flex-col gap-2', className)}>
       <div className="flex items-center justify-between">
         <Label>{label}</Label>
         <Input
@@ -92,11 +92,11 @@ export function StockSlider({
           onChange={handleSliderChange}
           disabled={disabled}
           className={cn(
-            "w-full h-2 appearance-none absolute top-0 left-0",
-            "bg-transparent focus:outline-none",
-            "[&::-webkit-slider-thumb]:appearance-none [&::-webkit-slider-thumb]:w-4 [&::-webkit-slider-thumb]:h-4 [&::-webkit-slider-thumb]:rounded-full [&::-webkit-slider-thumb]:bg-white [&::-webkit-slider-thumb]:border [&::-webkit-slider-thumb]:border-primary [&::-webkit-slider-thumb]:cursor-pointer [&::-webkit-slider-thumb]:shadow",
-            "[&::-moz-range-thumb]:appearance-none [&::-moz-range-thumb]:w-4 [&::-moz-range-thumb]:h-4 [&::-moz-range-thumb]:rounded-full [&::-moz-range-thumb]:bg-white [&::-moz-range-thumb]:border [&::-moz-range-thumb]:border-primary [&::-moz-range-thumb]:cursor-pointer [&::-moz-range-thumb]:shadow",
-            { "opacity-50 cursor-not-allowed": disabled }
+            'w-full h-2 appearance-none absolute top-0 left-0',
+            'bg-transparent focus:outline-none',
+            '[&::-webkit-slider-thumb]:appearance-none [&::-webkit-slider-thumb]:w-4 [&::-webkit-slider-thumb]:h-4 [&::-webkit-slider-thumb]:rounded-full [&::-webkit-slider-thumb]:bg-white [&::-webkit-slider-thumb]:border [&::-webkit-slider-thumb]:border-primary [&::-webkit-slider-thumb]:cursor-pointer [&::-webkit-slider-thumb]:shadow',
+            '[&::-moz-range-thumb]:appearance-none [&::-moz-range-thumb]:w-4 [&::-moz-range-thumb]:h-4 [&::-moz-range-thumb]:rounded-full [&::-moz-range-thumb]:bg-white [&::-moz-range-thumb]:border [&::-moz-range-thumb]:border-primary [&::-moz-range-thumb]:cursor-pointer [&::-moz-range-thumb]:shadow',
+            { 'opacity-50 cursor-not-allowed': disabled }
           )}
         />
       </div>
